Module chain_evm::machine

The Virtual Machine

Abstractions for the EVM runtime.

This module contains types that abstract away implementation details from the evm crate, specially those involving node and stack configurations, and runtime execution.

Handler <- EVM Context Handler

StackState<’config>

Modules

test

Structs

AccessListItem

Export EVM types

Environment

EVM Environment parameters needed for execution.

Log

VirtualMachine

Enums

Config

EVM Configuration parameters needed for execution.

Error

ExitError

Exit error reason.

Traits

EvmState

Functions

estimate_transact_call

Estimate a CALL transaction

estimate_transact_create

Estimate a CREATE transaction

estimate_transact_create2

Estimate a CREATE2 transaction

execute_transact_call

Execute a CALL transaction

execute_transact_create

Execute a CREATE transaction

execute_transact_create2

Execute a CREATE2 transaction

generate_address_create

generate_address_create2

Type Definitions

AccessList

Export EVM types

Address

An address of an EVM account.

BlockBaseFeePerGas

A block’s base fee per gas.

BlockCoinBase

A block’s coinbase

BlockDifficulty

A block’s difficulty.

BlockGasLimit

A block’s gas limit.

BlockHash

A block hash.

BlockHashes

A block hash.

BlockNumber

A block’s number.

BlockTimestamp

A block’s timestamp.

ChainId

A block’s chain ID.

Gas

Gas quantity integer for EVM operations.

GasLimit

Gas limit for EVM operations.

GasPrice

Gas price integer for EVM operations.

RuntimeContext

The context of the EVM runtime