Module chain_evm::state

This module contains a simple representation of all EVM-related data based on immutable data structures. It is designed to be used with the ledger implementation from chain-impl-mockchain in Jormungandr multiverse of ledgers.

Structs

Account

A represantation of an EVM account.

AccountState

A representation of an EVM account state.

Balance

Ethereum account balance which uses the least 64 significant bits of the U256 type.

LogsState

Trie

An immutable structure to represent any of EVM tries.

Enums

Error

Definition for state-related errors.

Type Definitions

AccountTrie

In-memory representation of all accounts.

ByteCode

Smart-contract bytecode, such as the one compiled from Solidity code, for example.

Key

Representation of a storage key. Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

Storage

In-memory representation of account storage.

Value

Representation of a storage key. Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.