Type Definition chain_evm::state::AccountTrie

pub type AccountTrie = Trie<Address, Account>;

In-memory representation of all accounts.

Implementations

impl AccountTrie

pub fn modify_account<F>(self, address: Address, f: F) -> Selfwhere
F: FnOnce(Account) -> Option<Account>,

Modify account

If the element is not present, the closure F is apllied to the Default::default() value, otherwise the closure F is applied to the found element. If the closure returns None, then the key is deleted